Split django test files
